% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_colors.R
\name{ceris_diverge_palette}
\alias{ceris_diverge_palette}
\title{CERIS diverging color palette for correlation heatmaps}
\usage{
ceris_diverge_palette(n = 26L)
}
\arguments{
\item{n}{Number of colors (default 26)}
}
\value{
Character vector of hex colors
}
\description{
CERIS diverging color palette for correlation heatmaps
}
\examples{
pal <- ceris_diverge_palette(11)
length(pal)
}
