% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_colors.R
\name{ceris_env_palette}
\alias{ceris_env_palette}
\title{Environment color palette}
\usage{
ceris_env_palette(n_envs)
}
\arguments{
\item{n_envs}{Number of environments}
}
\value{
Character vector of hex colors with alpha
}
\description{
Environment color palette
}
\examples{
colors <- ceris_env_palette(7)
length(colors)
}
