% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_colors.R
\name{ceris_grey_alpha}
\alias{ceris_grey_alpha}
\title{Semi-transparent grey for background data}
\usage{
ceris_grey_alpha()
}
\value{
A single color string
}
\description{
Semi-transparent grey for background data
}
\keyword{internal}
