% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceris_identify_best.R
\name{ceris_identify_best}
\alias{ceris_identify_best}
\title{Identify the Best Environmental Window from CERIS Results}
\usage{
ceris_identify_best(ceris_result, params, min_window = 7)
}
\arguments{
\item{ceris_result}{Data.frame from \code{run_CERIS}}

\item{params}{Character vector of parameter names}

\item{min_window}{Minimum window size in days (default 7)}
}
\value{
A list with components:
  \item{param_name}{Name of the best environmental parameter}
  \item{dap_start}{Start day of the best window}
  \item{dap_end}{End day of the best window}
  \item{correlation}{Correlation value at the best window}
  \item{neg_log_p}{-log10(p-value) at the best window}
}
\description{
Finds the window and parameter with the maximum absolute correlation
from a CERIS search result.
}
\examples{
\donttest{
d <- load_crop_data("sorghum")
exp_trait <- prepare_trait_data(d$traits, "FTdap")
env_mean_trait <- compute_env_means(exp_trait, d$env_meta)
params <- c("DL", "GDD", "PTT", "PTR", "PTS")
result <- run_CERIS(env_mean_trait, d$env_params, params, max_days = 80)
best <- ceris_identify_best(result, params)
best$param_name
best$dap_start
best$dap_end
}
}
