% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_colors.R
\name{ceris_poly_alpha}
\alias{ceris_poly_alpha}
\title{Semi-transparent violet for IQR polygons}
\usage{
ceris_poly_alpha()
}
\value{
A single color string
}
\description{
Semi-transparent violet for IQR polygons
}
\keyword{internal}
