% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{compute_window_params}
\alias{compute_window_params}
\title{Compute environmental covariate values for a window}
\usage{
compute_window_params(env_mean_trait, env_params, dap_start, dap_end, params)
}
\arguments{
\item{env_mean_trait}{Data.frame with env_code column}

\item{env_params}{Data.frame with env_code, DAP, and parameter columns}

\item{dap_start}{Start day of window}

\item{dap_end}{End day of window}

\item{params}{Character vector of parameter column names}
}
\value{
Data.frame env_mean_trait with added kPara column (first param) and
  a \code{window_params} attribute containing all parameter averages
}
\description{
For each environment, calculates the mean of each environmental parameter
over a specified DAP window.
}
\examples{
d <- load_crop_data("sorghum")
exp_trait <- prepare_trait_data(d$traits, "FTdap")
env_mean_trait <- compute_env_means(exp_trait, d$env_meta)
params <- c("DL", "GDD", "PTT", "PTR", "PTS")
env_mean_trait <- compute_window_params(env_mean_trait, d$env_params, 20, 60, params)
head(env_mean_trait[, c("env_code", "meanY", "kPara")])
}
