% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crop_info}
\alias{crop_info}
\title{Crop information summary}
\format{
Data frame with columns: crop, n_envs, traits, default_trait, has_genotype, env_params
}
\usage{
crop_info
}
\description{
Metadata about available crop datasets.
}
\keyword{datasets}
