% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_next_year}
\alias{forecast_next_year}
\title{Forecast Phenotypes for New Environments}
\usage{
forecast_next_year(exp_trait, env_mean_trait, trn_env)
}
\arguments{
\item{exp_trait}{Data.frame with line_code, env_code, Yobs}

\item{env_mean_trait}{Data.frame with env_code, meanY, kPara}

\item{trn_env}{Character vector of training environment codes}
}
\value{
Data.frame with columns: env_code, line_code, Prd_trait_mean,
  Prd_trait_kPara, Obs_trait, Line_mean
}
\description{
Predicts phenotypes for a set of environments using models trained on
a different set of environments (e.g., year-to-year prediction).
}
\examples{
\donttest{
d <- load_crop_data("sorghum")
exp_trait <- prepare_trait_data(d$traits, "FTdap")
env_mean_trait <- compute_env_means(exp_trait, d$env_meta)
params <- c("DL", "GDD", "PTT", "PTR", "PTS")
env_mean_trait <- compute_window_params(env_mean_trait, d$env_params, 20, 60, params)
trn_env <- env_mean_trait$env_code[1:5]
fc_result <- forecast_next_year(exp_trait, env_mean_trait, trn_env)
head(fc_result)
}
}
