% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{load_crop_data}
\alias{load_crop_data}
\title{Load a crop dataset}
\usage{
load_crop_data(crop)
}
\arguments{
\item{crop}{One of "sorghum", "maize", "rice", "oat"}
}
\value{
A list with components: traits, env_meta, env_params, genotype (NULL if unavailable)
}
\description{
Convenience function to load all components of a crop dataset.
}
\examples{
data_list <- load_crop_data("sorghum")
str(data_list$traits)
}
