% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maize_env_meta}
\alias{maize_env_meta}
\title{Maize environment metadata}
\format{
Data frame with columns: env_code, lat, lon, PlantingDate, TrialYear, env_note
}
\usage{
maize_env_meta
}
\description{
Maize environment metadata
}
\keyword{datasets}
