% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maize_env_params}
\alias{maize_env_params}
\title{Maize daily environmental parameters}
\format{
Data frame with columns: env_code, DAP, TMAX, TMIN, DL, GDD, PTT, PTR, PTS
}
\usage{
maize_env_params
}
\description{
Maize daily environmental parameters
}
\keyword{datasets}
