% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maize_genotype}
\alias{maize_genotype}
\title{Maize SNP genotype matrix}
\format{
Integer matrix with line_code as rownames
}
\usage{
maize_genotype
}
\description{
Maize SNP genotype matrix
}
\keyword{datasets}
