% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maize_traits}
\alias{maize_traits}
\title{Maize phenotypic trait records}
\format{
Data frame with columns: line_code, env_code, FT, PH
}
\usage{
maize_traits
}
\description{
Phenotypic data from 10 maize multi-environment trials.
}
\keyword{datasets}
