% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oat_env_meta}
\alias{oat_env_meta}
\title{Oat environment metadata}
\format{
Data frame with columns: env_code, lat, lon, PlantingDate, TrialYear, Location, env_note
}
\usage{
oat_env_meta
}
\description{
Oat environment metadata
}
\keyword{datasets}
