% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oat_env_params}
\alias{oat_env_params}
\title{Oat daily environmental parameters}
\format{
Data frame with columns: env_code, DAP, TMAX, TMIN, DL, GDD, PTT, PTR, PTD1, PTD2, PTS
}
\usage{
oat_env_params
}
\description{
Oat daily environmental parameters
}
\keyword{datasets}
