% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oat_traits}
\alias{oat_traits}
\title{Oat phenotypic trait records}
\format{
Data frame with columns: line_code, env_code, FTdap, PH, GY
}
\usage{
oat_traits
}
\description{
Phenotypic data from 13 oat multi-environment trials.
}
\keyword{datasets}
