% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_jra.R
\name{plot_jra}
\alias{plot_jra}
\title{Plot JRA Results}
\usage{
plot_jra(exp_trait, env_mean_trait, jra_result, trait = "Trait")
}
\arguments{
\item{exp_trait}{Data.frame with line_code, env_code, Yobs}

\item{env_mean_trait}{Data.frame with meanY}

\item{jra_result}{Data.frame from \code{jra_model} with Intcp, Slope_mean, R2_mean}

\item{trait}{Character; trait name for axis labels}
}
\value{
A patchwork object
}
\description{
Two-panel plot: (A) JRA regression lines overlaid on data, (B) histogram
of R-squared values.
}
\examples{
d <- load_crop_data("sorghum")
exp_trait <- prepare_trait_data(d$traits, "FTdap")
env_mean_trait <- compute_env_means(exp_trait, d$env_meta)
line_by_env <- prepare_line_by_env(exp_trait, env_mean_trait)
jra_result <- jra_model(line_by_env, env_mean_trait)
plot_jra(exp_trait, env_mean_trait, jra_result, trait = "FTdap")
}
