% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reaction_norm.R
\name{plot_reaction_norm}
\alias{plot_reaction_norm}
\title{Plot Reaction Norms}
\usage{
plot_reaction_norm(exp_trait, env_mean_trait, trait = "Trait")
}
\arguments{
\item{exp_trait}{Data.frame with line_code, env_code, Yobs}

\item{env_mean_trait}{Data.frame with env_code, meanY}

\item{trait}{Character; trait name}
}
\value{
A patchwork object
}
\description{
Four-panel reaction norm visualization: (A) raw data by arbitrary env order,
(B) by env mean order, (C) continuous env mean axis, (D) fitted regression lines.
}
\examples{
d <- load_crop_data("sorghum")
exp_trait <- prepare_trait_data(d$traits, "FTdap")
env_mean_trait <- compute_env_means(exp_trait, d$env_meta)
plot_reaction_norm(exp_trait, env_mean_trait, trait = "FTdap")
}
