% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_rrblup.R
\name{pred_rrblup}
\alias{pred_rrblup}
\title{Genomic Prediction using rrBLUP}
\usage{
pred_rrblup(Y_matrix, X_matrix, prd_idx, n)
}
\arguments{
\item{Y_matrix}{Data.frame with ID_code as first column and trait columns}

\item{X_matrix}{Genotype matrix with line codes as rownames}

\item{prd_idx}{Integer vector of row indices for the validation set}

\item{n}{Iteration number (for tracking)}
}
\value{
Data.frame with observed and predicted values for each trait
}
\description{
Predicts phenotypic values for a validation set using ridge regression BLUP.
}
\examples{
\donttest{
d <- load_crop_data("sorghum")
exp_trait <- prepare_trait_data(d$traits, "FTdap")
env_mean_trait <- compute_env_means(exp_trait, d$env_meta)
params <- c("DL", "GDD", "PTT", "PTR", "PTS")
env_mean_trait <- compute_window_params(env_mean_trait, d$env_params, 20, 60, params)
lm_ab <- slope_intercept(exp_trait, env_mean_trait, type = "kPara")
SNPs <- prepare_genotype(d$genotype, unique(exp_trait$line_code))
ab_df <- lm_ab[, c("line_code", "Intcp_para", "Slope_para")]
prd_result <- pred_rrblup(ab_df, SNPs, prd_idx = 1:5, n = 1)
head(prd_result)
}
}
