% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_line_by_env}
\alias{prepare_line_by_env}
\title{Prepare line-by-environment matrix}
\usage{
prepare_line_by_env(exp_trait, env_mean_trait)
}
\arguments{
\item{exp_trait}{Data.frame from \code{prepare_trait_data}}

\item{env_mean_trait}{Data.frame from \code{compute_env_means}}
}
\value{
Wide data.frame with line_code + environment columns
}
\description{
Creates a wide-format data.frame with line_code as first column and
environments as subsequent columns, ordered by environmental mean.
}
\examples{
d <- load_crop_data("sorghum")
exp_trait <- prepare_trait_data(d$traits, "FTdap")
env_mean_trait <- compute_env_means(exp_trait, d$env_meta)
line_by_env <- prepare_line_by_env(exp_trait, env_mean_trait)
head(line_by_env[, 1:4])
}
