% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rice_env_meta}
\alias{rice_env_meta}
\title{Rice environment metadata}
\format{
Data frame with columns: env_code, lat, lon, PlantingDate, TrialYear, Location, env_note
}
\usage{
rice_env_meta
}
\description{
Rice environment metadata
}
\keyword{datasets}
