% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rice_env_params}
\alias{rice_env_params}
\title{Rice daily environmental parameters}
\format{
Data frame with columns: env_code, DAP, TMAX, TMIN, DL, GDD, PTT, PTR, PTS
}
\usage{
rice_env_params
}
\description{
Rice daily environmental parameters
}
\keyword{datasets}
