% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rice_genotype}
\alias{rice_genotype}
\title{Rice SNP genotype matrix}
\format{
Integer matrix with line_code as rownames
}
\usage{
rice_genotype
}
\description{
Rice SNP genotype matrix
}
\keyword{datasets}
