% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rice_traits}
\alias{rice_traits}
\title{Rice phenotypic trait records}
\format{
Data frame with columns: line_code, env_code, FTdap, FTgdd
}
\usage{
rice_traits
}
\description{
Phenotypic data from 9 rice multi-environment trials.
}
\keyword{datasets}
