% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the CERIS Shiny Application}
\usage{
run_app(...)
}
\arguments{
\item{...}{Arguments passed to \code{shiny::runApp}}
}
\description{
Launch the CERIS Shiny Application
}
\examples{
\dontrun{
run_app()
}
}
