% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_intercept.R
\name{slope_intercept}
\alias{slope_intercept}
\title{Calculate Reaction Norm Slopes and Intercepts}
\usage{
slope_intercept(exp_trait, env_mean_trait, type = "kPara")
}
\arguments{
\item{exp_trait}{Data.frame from \code{prepare_trait_data} with line_code, env_code, Yobs}

\item{env_mean_trait}{Data.frame from \code{compute_env_means} with meanY and kPara columns}

\item{type}{Character; one of "kPara" (default), "mean", or "both"}
}
\value{
Data.frame with line_code and slope/intercept columns depending on type:
  \itemize{
    \item "kPara": Intcp_para_adj, Intcp_para, Slope_para, R2_para
    \item "mean": Intcp_mean, Slope_mean
    \item "both": all columns above
  }
}
\description{
Fits linear regressions of individual line phenotypes against an
environmental covariate (kPara) and/or the environmental mean.
}
\examples{
d <- load_crop_data("sorghum")
exp_trait <- prepare_trait_data(d$traits, "FTdap")
env_mean_trait <- compute_env_means(exp_trait, d$env_meta)
params <- c("DL", "GDD", "PTT", "PTR", "PTS")
env_mean_trait <- compute_window_params(env_mean_trait, d$env_params, 20, 60, params)
si <- slope_intercept(exp_trait, env_mean_trait, type = "kPara")
head(si)
}
