% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sorghum_env_meta}
\alias{sorghum_env_meta}
\title{Sorghum environment metadata}
\format{
Data frame with columns: env_notes, env_code, lat, lon, PlantingDate, TrialYear, Location
}
\usage{
sorghum_env_meta
}
\description{
Location and planting information for sorghum trial environments.
}
\keyword{datasets}
