% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sorghum_env_params}
\alias{sorghum_env_params}
\title{Sorghum daily environmental parameters}
\format{
Data frame with columns: env_code, DAP, DL, GDD, PTT, PTR, PTS
}
\usage{
sorghum_env_params
}
\description{
Daily environmental covariates for sorghum trial environments.
}
\keyword{datasets}
