% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sorghum_genotype}
\alias{sorghum_genotype}
\title{Sorghum SNP genotype matrix}
\format{
Integer matrix with line_code as rownames and marker names as colnames
}
\usage{
sorghum_genotype
}
\description{
SNP marker data for sorghum lines.
}
\keyword{datasets}
