% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sorghum_traits}
\alias{sorghum_traits}
\title{Sorghum phenotypic trait records}
\format{
Data frame with columns: env_code, pop_code, line_code, FTdap, FTgdd, env_note
}
\source{
Li & Guo CERIS_PAG repository
}
\usage{
sorghum_traits
}
\description{
Phenotypic data from 7 sorghum multi-environment trials.
}
\keyword{datasets}
